+++ definitions/interfaces/module-metadata.interface.ts
@@ -1,7 +1,30 @@
 import {ModuleSubCollection} from './module-sub-collection.interface';
 
 export interface ModuleMetadata {
+
+  /**
+   * Defines if a module has attached files.
+   * This defaults to true and only setting `attachedFiles.contains = false`
+   * would prevent the delete trigger from attempting
+   * to delete attached files
+   */
+  attachedFiles?: {
+    containes?: boolean;
+
+    /**
+     * Folder to search for files to delete in.
+     * Defaults to "/". Can also contain dynamic {{documentId}} and
+     * {{moduleId}} macros.
+     */
+    prefix?: string;
+  };
+
   subCollections?: ModuleSubCollection[];
+
+  /**
+   * When set to true attempts to delete an auth user
+   * with the same id of the deleted document
+   */
   deletedAuthUser?: boolean;
 
   /**
@@ -31,5 +54,10 @@ export interface ModuleMetadata {
    * Confirmation dialog to confirm page exit if form is edited
    */
   confirmExitOnTouched?: boolean;
+
+  /**
+   * Generates history entries for the specified collection
+   */
+  history?: boolean;
   [key: string]: any;
 }