+++ definitions/modules/modules.ts
@@ -3,6 +3,7 @@ import {SENT_EMAILS_MODULE} from './emails/sent-emails.module';
 import {FORMS_MODULE} from './forms.module';
 import {PAGES_MODULE} from './pages/pages.module';
 import {ROLES_MODULE} from './roles.module';
+import {USER_HISTORY_MODULE} from './user-history.module';
 import {USER_INVITES_MODULE} from './user-invites.module';
 import {USERS_MODULE} from './users.module';
 import {LAYOUT_MODULE} from './website/layout.module';
@@ -18,10 +19,12 @@ export const MODULES = [
   AUTOMATIC_EMAILS_MODULE,
   SENT_EMAILS_MODULE,
 
+  USER_HISTORY_MODULE,
+
   /**
    * Website
    */
-   PAGES_MODULE,
-   FORMS_MODULE,
-   LAYOUT_MODULE,
+  PAGES_MODULE,
+  FORMS_MODULE,
+  LAYOUT_MODULE,
 ];